Move all settings code from `util-server.js` into `settings.js`
